Remove all test code
